// src/util/interfaces/ParsedObjects/gamemodes/ParsedPit.ts
export interface ParsedPit {
  experience: string;
  cash: string;
  playtime: string;

  goldenApplesEaten: string;
  jumpsIntoPit: string;
  ingotsPickedUp: string;

  kills: string;
  deaths: string;
  kdr: string;

  bowHits: string;
  bowShots: string;
  bowAccuracy: string;

  highestKillstreak: string;
  damageDealt: string;
  damageReceived: string;
}